pyproject.toml
[project]
name = "codereview-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.84.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
]
